renaming translations file
